requirements.txt
python-telegram-bot==21.3
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

